+++ packages/expo/src/launch/withExpoRoot.types.ts
@@ -1,7 +1,6 @@
 export type InitialProps = {
   exp: {
     notification?: any;
-    errorRecovery?: any;
     manifestString?: string;
     [key: string]: any;
   };